refactor: split rival tab view, viewmodel, and models
